pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"

dependencies = [
    "authlib~=1.6.5",
    "flask~=3.1.1",
    "python-dotenv~=1.0.0",
    "flask-login~=0.6.3",
    "flask-migrate~=4.0.7",
    "flask-sqlalchemy~=3.1.1",
    "flask-wtf~=1.2.2",
    "gunicorn~=23.0.0",
    "jinja2~=3.1.4",
    "matplotlib~=3.10.5",
    "mercadopago~=2.3.0",
    "openai~=1.97.1",
    "openpyxl~=3.1.5",
    "pillow~=11.3.0",
    "psycopg[binary]>=3.2,<3.3",   # reemplaza psycopg2-binary por psycopg v3
    "pypdf2~=3.0.1",
    "reportlab~=4.4.2",
    "requests~=2.32.3",
    "sqlalchemy~=2.0.41",
    "werkzeug~=3.1.3",
    "weasyprint~=66.0",
    "xlsxwriter~=3.2.5",
]

[project.optional-dependencies]
dev = [
    "pytest~=8.0.0",
    "pytest-flask~=1.3.0",
    "pytest-cov~=5.0.0",
    "pytest-mock~=3.14.0",
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=.",
    "--cov-report=html",
    "--cov-report=term-missing",
]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["."]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/.venv/*",
    "*/venv/*",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

